return corr matrix
