+++ macrosynergy/panel/__init__.py
@@ -8,7 +8,7 @@ from .make_blacklist import make_blacklist
 from .make_relative_value import make_relative_value
 from .make_zn_scores import make_zn_scores
 from .panel_calculator import panel_calculator
-from .view_correlations import correl_matrix
+from .correlation import correl_matrix
 from .view_grades import heatmap_grades
 from .view_ranges import view_ranges
 from .view_timelines import view_timelines
